BackButtonComponent- added back button on registration form page
